update get mime type

+++ app/src/main/java/com/example/videolightcompressor/ui/gallery/VideoDetailFragment.kt
@@ -1,20 +1,22 @@
 package com.example.videolightcompressor.ui.gallery
 
+import android.Manifest
 import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import com.example.videolightcompressor.R
+import com.tedpark.tedpermission.rx2.TedRx2Permission
 import gun0912.tedimagepicker.base.BaseBottomSheetFragment
-import com.example.videolightcompressor.event.DeleteVideoEvent
-import com.example.videolightcompressor.extensions.showAlertDialog
-import com.example.videolightcompressor.extensions.toast
-import gun0912.tedimagepicker.extenstion.extractVideoInfo
-import gun0912.tedimagepicker.extenstion.getFileFromUri
+import gun0912.tedimagepicker.event.DeleteVideoEvent
+import gun0912.tedimagepicker.extenstion.*
+import io.reactivex.disposables.CompositeDisposable
 import kotlinx.android.synthetic.main.fragment_video_detail.*
 import org.greenrobot.eventbus.EventBus
 
 class VideoDetailFragment :
     BaseBottomSheetFragment() {
+
+    private val disposable = CompositeDisposable()
 
     override fun layoutResId(): Int = R.layout.fragment_video_detail
 
@@ -53,16 +55,27 @@ class VideoDetailFragment :
     override fun onDestroyView() {
         super.onDestroyView()
         videoView.releasePlayer()
+        disposable.clear()
     }
 
     private fun deleteVideo(uri: Uri) {
-        val file = uri.getFileFromUri(requireContext())
-        if (file.delete()) {
-            EventBus.getDefault().post(DeleteVideoEvent())
-            dismiss()
-        } else {
-            context?.toast(R.string.video_deleted_msg_failed)
-        }
+        TedRx2Permission.with(context)
+            .setPermissions(Manifest.permission.WRITE_EXTERNAL_STORAGE)
+            .request()
+            .subscribe({ permissionResult ->
+                if (permissionResult.isGranted) {
+                    val file = uri.getFileFromUri(requireContext())
+                    if (file.delete()) {
+                        EventBus.getDefault().post(DeleteVideoEvent())
+                        dismiss()
+                    } else {
+                        context?.toast(R.string.video_deleted_msg_failed)
+                    }
+                } else {
+                    context?.toast(getString(R.string.permission_denied))
+                }
+            }, { throwable -> context?.toast(throwable.localizedMessage ?: "") })
+            .addTo(disposable)
     }
 
     companion object {
